# perf: parallelize slow formatting steps