With the more C++ and CMake files being added, the check-style step is
getting slower. The cmake-format step is the slowest, and was taking
about 45 seconds on my laptop. clang-format was taking about 23 seconds.

This change parallelizes the executions of these steps with `xargs -P`.
The parallelism and number of arguments values were determined
experimentally on my machine.

With this change, the cmake-format step is about 11 seconds, and the
clang-format step is about 4 seconds.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/googleapis/google-cloud-cpp/3912)
<!-- Reviewable:end -->